Change & Clean Code v8.

+++ src/main/java/ir/mostashar/model/client/service/ClientService.java
@@ -264,10 +264,10 @@ public class ClientService {
         smsService.sendSms(2, mobileNumber, code);
     }
 
-    public void minusScore(String clientId, int score) {
+    public void minusScore(String clientId, double score) {
         Optional<Client> client = clientRepo.findByUid(UUID.fromString(clientId));
         if (client.isPresent()) {
-            int totalScore = client.get().getScore();
+            double totalScore = client.get().getScore();
             totalScore -= score;
             client.get().setScore(totalScore);
             clientRepo.save(client.get());
